move inline off of formatblot and remove
